api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
singlestoredb==1.7.0
boto3>=1.35.0
botocore>=1.35.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
websockets==14.1
